get google and git auth working again
