+++ package.json
@@ -51,10 +51,10 @@
         "@tiptap/starter-kit": "^2.0.2",
         "@types/bcryptjs": "^2.4.2",
         "@types/node": "18.15.11",
-        "@types/react": "18.0.34",
+        "@types/react": "18.0.35",
         "@types/react-dom": "18.0.11",
         "autoprefixer": "^10.4.14",
-        "aws-sdk": "^2.1354.0",
+        "aws-sdk": "^2.1356.0",
         "axios": "^1.3.5",
         "bcryptjs": "^2.4.3",
         "classnames": "^2.3.2",
@@ -62,7 +62,7 @@
         "cookie": "^0.5.0",
         "dayjs": "^1.11.7",
         "embla-carousel-react": "^7.1.0",
-        "eslint-config-next": "13.1.5",
+        "eslint-config-next": "13.3.0",
         "focus-visible": "^5.2.0",
         "next": "12.3.x",
         "next-auth": "^4.22.0",
@@ -70,7 +70,6 @@
         "nextjs-backend-helpers": "^0.0.18-8",
         "openai": "^3.2.1",
         "postcss-focus-visible": "^8.0.2",
-        "prisma": "^4.12.0",
         "query-string": "^8.1.0",
         "react": "18.2.0",
         "react-dom": "18.2.0",
@@ -87,7 +86,7 @@
         "@types/uuid": "^9.0.1",
         "@typescript-eslint/eslint-plugin": "^5.58.0",
         "@typescript-eslint/parser": "^5.58.0",
-        "autoprefixer": "^10.4.13",
+        "autoprefixer": "^10.4.14",
         "cross-env": "^7.0.3",
         "dotenv": "^16.0.3",
         "dotenv-cli": "^7.2.1",
@@ -98,9 +97,10 @@
         "postcss": "^8.4.21",
         "prettier": "^2.8.7",
         "prettier-plugin-tailwindcss": "^0.2.7",
+        "prisma": "^4.12.0",
         "readline-sync": "^1.4.10",
         "rimraf": "^5.0.0",
-        "tailwindcss": "^3.2.7",
+        "tailwindcss": "^3.3.1",
         "typescript": "5.0.4",
         "uuid": "^9.0.0"
     }
