+++ src/components/auth/Providers.tsx
@@ -1,93 +1,8 @@
-import React, { useState } from 'react';
-import { Button } from '@/components/buttons/Button';
-import { signIn } from 'next-auth/react';
-import { DiGithubBadge, DiGoogleDrive } from 'react-icons/di';
 import { CommonProviderOptions } from 'next-auth/providers';
 import { env } from '@/env/client.mjs';
-import { TextField } from '@/components/landing/Fields';
-import Divider from '../dividers/Divider';
+import { GithubProvider, GooogleProvider } from './CloudProviders';
+import { CredentialsProvider } from './CredentialsProvider';
 
-function handleCloudAuthSignin(provider: CommonProviderOptions) {
-    return async () => {
-        await signIn(provider.id, {
-            redirect: true,
-            callbackUrl: "http://localhost:3000/portal"
-            // callbackUrl: `${window.location.origin}/portal`
-        });
-    };
-}
-
-function handlePasswordSignin(provider: CommonProviderOptions, emailaddress: string, password: string) {
-    return async () => {
-        await signIn(provider.id, {
-            emailaddress,
-            password,
-            redirect: true,
-            callbackUrl: `${window.location.origin}/portal`
-        });
-    };
-}
-
-function GithubProvider({ provider }: { provider: CommonProviderOptions }) {
-    return (
-        <div className="mb-12" key={provider.name}>
-            <Button variant="solid" color="blue" className="w-full" onClick={handleCloudAuthSignin(provider)}>
-                <span>
-                    <DiGithubBadge className="-mt-1 inline" size={'2rem'} /> Sign in with {provider.name} <span aria-hidden="true">&rarr;</span>
-                </span>
-            </Button>
-        </div>
-    );
-}
-
-function GooogleProvider({ provider }: { provider: CommonProviderOptions }) {
-    return (
-        <div className="mb-12" key={provider.name}>
-            <Button variant="solid" color="blue" className="w-full" onClick={handleCloudAuthSignin(provider)}>
-                <span>
-                    <DiGoogleDrive className="-mt-1 inline" size={'2rem'} /> Sign in with {provider.name} <span aria-hidden="true">&rarr;</span>
-                </span>
-            </Button>
-        </div>
-    );
-}
-
-
-function CredentialsProvider({ provider }: { provider: CommonProviderOptions }) {
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-
-    const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        setEmail(event.target.value);
-    };
-
-    const handlePassChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        setPassword(event?.target.value);
-    };
-
-    return (
-        <div className="mt-10 grid grid-cols-1 gap-y-8">
-            <p className="font-extrabold">Sign with {provider.name} </p>
-            <TextField
-                onChange={handleEmailChange}
-                label="Email address"
-                id="usernameoremail"
-                name="usernameoremail"
-                type="text"
-                autoComplete="email"
-                required
-            />
-            <TextField onChange={handlePassChange} label="Password" id="password" name="password" type="password" autoComplete="current-password" required />
-            <div className="mb-12">
-                <Button type="submit" variant="solid" color="blue" className="w-full" onClick={handlePasswordSignin(provider, email, password)}>
-                    <span>
-                        Sign in <span aria-hidden="true">&rarr;</span>
-                    </span>
-                </Button>
-            </div>
-        </div>
-    );
-}
 
 export function Providers({ providers = [] }: { providers: CommonProviderOptions[] }) {
     return (
